修复预览显示的问题，更新README，祝大家新年快乐，新的一年事业有成

+++ library/src/main/java/com/brook/app/android/supportlibrary/util/Util.java
@@ -180,8 +180,8 @@ public class Util {
      * @param spVal
      * @return
      */
-    public static int sp2px(float spVal) {
+    public static float sp2px(float spVal) {
         final float fontScale = getContext().getResources().getDisplayMetrics().scaledDensity;
-        return (int) (spVal * fontScale + 0.5f);
+        return (spVal * fontScale + 0.5f);
     }
 }
